Rename extract_slices.py to do all preprocessing
